Improve CI workflow efficiency and fix linting issues
CI workflow improvements:
- Split lint job into separate fmt and clippy jobs for faster feedback
- Replace manual cargo cache with Swatinem/rust-cache for better deduplication
- Remove debug builds (only build release for artifacts)
- Add concurrency groups to cancel stale runs on new pushes
- Add paths-ignore to skip CI for docs-only changes
- Add timeout-minutes to all jobs to prevent hung workflows
- Use matrix include for cleaner artifact upload configuration
- Add ci-success summary job for branch protection rules

Code fixes:
- Fix rustfmt formatting issues across source files
- Use is_multiple_of() instead of manual modulo check
- Use strip_prefix() instead of manual string slicing
- Remove unnecessary clone on Copy type
- Remove unnecessary i64 cast
- Use div_ceil() instead of manual implementation
- Handle unused Result from AttachConsole
- Add #[allow(dead_code)] for planned but unused API methods
- Fix conditional import of Arc (Linux-only)

+++ crates/core/src/screen/time_sync.rs
@@ -293,7 +293,7 @@ mod tests {
             let client_time = 1000 + i as i64 * 100;
             let server_receive_time = client_time + 100 + rtt / 2;
             let server_send_time = server_receive_time + 1;
-            let response_received_time = client_time + *rtt as i64 + 1;
+            let response_received_time = client_time + *rtt + 1;
 
             sync.process_response(
                 client_time,
